Screencast: Fix screencast video cannot playback
Fix cannot play the screencast video after reboot the device.
stopRecorder is not called when system reboot/shutdown while
screencast is still recording in the background services,
since onDestory in Service does not get called in this case.
Add a ACTION_SHUTDOWN intent to listen to this event and handling
this gratefully.

FEIJ-404

Change-Id: If26f568841eb681e5e30a00a68de2fbe9adacc4a